src/microservices/matches-service/requirements.txt
fastapi
uvicorn
aiokafka
psycopg2-binary
sqlalchemy
prometheus-fastapi-instrumentator
python-logging-loki
